The course model has been updated to have the credit field as a double.

+++ lib/models/course_model.dart
@@ -13,7 +13,7 @@ class Course {
   }) : courseId = const Uuid().v1();
 
   String title;
-  int credits;
+  double credits;
   double totalHourCount;
   String courseId;
